⚠️READ: Moved to the "common" package all the business logic of the dependencies downloader. Now it's mandatory to implement the common package. A solution for this must be found
