+++ app/build.gradle
@@ -49,9 +49,10 @@ dependencies {
     implementation project(":library")
     implementation project(":ffmpeg")
     implementation project(":aria2c")
+    implementation project(":common")
 
     implementation "androidx.appcompat:appcompat:$appCompatVer"
-    implementation "androidx.constraintlayout:constraintlayout:1.1.3"
+    implementation "androidx.constraintlayout:constraintlayout:2.1.4"
     implementation "androidx.core:core-ktx:$coreKtxVer"
     testImplementation "junit:junit:$junitVer"
     androidTestImplementation "androidx.test.ext:junit:$androidJunitVer"
