+++ common/src/main/java/com/yausername/youtubedl_common/utils/Json.kt
@@ -1,0 +1,9 @@
+package com.yausername.youtubedl_common.utils
+
+import kotlinx.serialization.json.Json
+
+val json = Json {
+    ignoreUnknownKeys = true
+    encodeDefaults = true
+    coerceInputValues = true
+}
