+++ common/src/main/java/com/yausername/youtubedl_common/utils/files/FilesUtil.kt
@@ -1,4 +1,4 @@
-package com.yausername.youtubedl_android.util.files
+package com.yausername.youtubedl_common.utils.files
 
 import android.util.Log
 import java.io.File
