--- library/src/main/java/com/yausername/youtubedl_android/util/exceptions/MissingDependency.kt
@@ -1,7 +1,0 @@
-package com.yausername.youtubedl_android.util.exceptions
-
-class MissingDependency: Exception {
-    constructor(message: String?) : super(message) {}
-    constructor(message: String?, e: Throwable?) : super(message, e) {}
-    constructor(e: Throwable?) : super(e) {}
-}
